GADE_HAND_IN
In this upload I have not changed any code in comparison with the Unity project file, they are completely the same, i just had internet problems yesterday when i was trying to upload the Unity file, and this updated github local repository. i have tested my unity game and for some reason when all buildings are destroyed the game should end so say which team is victorious, so i have put an if statement to check if the buildings are destroyed and if ture it should end the game,for example; instead it detects that there are no Blue Factories but then it does not end the game. same for the red team. but when a team reaches the end goal of 100 kills the game then ends and resets for the next round. So honestly i just believe its a glitch, cause either then that my code works fine.