bug fix
This is the code with the bug fixed, dont  know if this will change anything but i just thought to upload the bug fixed code anyway.
The bug that i fixed was the one where i check if all "Blue Factories" or "Red Factories" are destroyed where it did not end the game if "true". All i done here was change the if statement to check if the array holding the "Factories" is <= 0. the code in the final uploaded Unity file checks if the array holding the "Factories" == null, since its empty, but it didn't work, so i guess here is the fixed code.